+++ packages/effects-core/src/math/value-getters/color-curve.ts
@@ -1,7 +1,8 @@
 import { Color } from '@galacean/effects-math/es/core/color';
 import type { BezierCurve } from './value-getter';
-import { ValueGetter, createValueGetter } from './value-getter';
+import { ValueGetter } from './value-getter';
 import type { spec } from '@galacean/effects-core';
+import { createValueGetter } from './value-getter-map';
 
 export class ColorCurve extends ValueGetter<Color> {
   private value = new Color();
