+++ packages/effects-core/src/math/value-getters/index.ts
@@ -1,3 +1,4 @@
 export * from './value-getter';
 export * from './color-curve';
 export * from './vector4-curve';
+export * from './value-getter-map';
