+++ packages/effects-core/src/math/value-getters/vector4-curve.ts
@@ -1,7 +1,8 @@
 import { Vector4 } from '@galacean/effects-math/es/core/vector4';
 import type { BezierCurve } from './value-getter';
-import { ValueGetter, createValueGetter } from './value-getter';
+import { ValueGetter } from './value-getter';
 import type { spec } from '@galacean/effects-core';
+import { createValueGetter } from './value-getter-map';
 
 export class Vector4Curve extends ValueGetter<Vector4> {
   private value = new Vector4();
